fixed float bug in uncond traintest

+++ DEwNF/utils/data_utils.py
@@ -163,10 +163,10 @@ def searchlog_unconditional_day_split(sup_df, unsup_df, obs_cols, batch_size, cu
     scaled_extra_obs = obs_scaler.transform(unsup_df.loc[:, obs_cols])
 
     # concatenate the train and extra data
-    scaled_train_data = torch.cat((torch.tensor(scaled_train_obs), torch.tensor(scaled_extra_obs)))
+    scaled_train_data = torch.cat((torch.tensor(scaled_train_obs), torch.tensor(scaled_extra_obs))).float()
 
     # Transform test data
-    scaled_test_data = torch.tensor(obs_scaler.transform(sup_df.loc[test_idx, obs_cols]))
+    scaled_test_data = torch.tensor(obs_scaler.transform(sup_df.loc[test_idx, obs_cols])).float()
 
     if cuda_exp:
         scaled_train_data = scaled_train_data.cuda()
